requirements1.txt
fastapi
uvicorn
sqlalchemy>=2
asyncpg
psycopg2-binary>=2.9.9
transformers
torch==2.2.2+cpu
langgraph
pytest
pytest-asyncio





accelerate==1.9.0
aiohappyeyeballs 
aiohttp 
aiosignal
annotated-types
anyio 
argon2-cffi 
argon2-cffi-bindings 
arrow 
asttokens 
async-lru 
async-timeout
attrs 
babel 
backports.asyncio.runner==1.2.0
beautifulsoup4 
bleach 
Bottleneck 
brotlicffi 
cached-property 
certifi 
cffi 
chardet==5.2.0
charset-normalizer 
click 
colorama 
comm 
datasets 
debugpy 
decorator 
defusedxml 
Deprecated 
diskcache 
dnspython 
email_validator 
exceptiongroup 
executing 
fastapi 
fastapi-cli 
fastjsonschema
filelock 
fqdn
frozenlist 
fsspec 
gmpy2
greenlet==3.2.3
h11 
h2 
hpack 
httpcore 
httptools 
httpx 
huggingface_hub 
hyperframe 
idna 
importlib_metadata 
importlib_resources
iniconfig==2.1.0
ipykernel
ipython
ipywidgets 
isoduration 
jedi 
Jinja2 
json5 
jsonpatch==1.33
jsonpointer 
jsonschema 
jsonschema-specifications 
jupyter-events 
jupyter-lsp 
jupyter_client 
jupyter_server
jupyter_server_terminals
jupyterlab 
jupyterlab_pygments
jupyterlab_server 
jupyterlab_widgets 
langchain-core==0.3.69
langgraph==0.5.3
langgraph-checkpoint==2.1.0
langgraph-prebuilt==0.5.2
langgraph-sdk==0.1.73
langsmith==0.4.6
llama_cpp_python 
markdown-it-py 
MarkupSafe 
matplotlib-inline 
mdurl 
mistune
mkl-service==2.4.0
mkl_fft 
mkl_random 
mpmath 
multidict
multiprocess
nbclient 
nbconvert 
nbformat
nest_asyncio 
networkx 
notebook_shim 
numexpr 
numpy 
opentelemetry-api 
orjson==3.11.0
ormsgpack==1.10.0
overrides 
packaging 
pandas 
pandocfilters 
parso 
pickleshare
pkgutil_resolve_name
platformdirs 
pluggy==1.6.0
prometheus_client 
prompt_toolkit 
propcache 
protobuf
psutil 
psycopg2-binary==2.9.10
pure_eval 
pyarrow 
pycparser 
pydantic 
pydantic-settings 
pydantic_core 
Pygments 
PySocks
pytest==8.4.1
pytest-asyncio==1.1.0
python-dateutil 
python-dotenv 
python-json-logger 
python-multipart 
pytz 
pywin32 
pywinpty
PyYAML 
pyzmq 
referencing 
regex 
requests 
requests-toolbelt==1.0.0
rfc3339_validator 
rfc3986-validator 
rich 
rich-toolkit 
rpds-py 
safetensors 
Send2Trash
shellingham 
six 
sniffio 
soupsieve 
SQLAlchemy==2.0.41
sse-starlette
stack_data
starlette 
starlette-context 
sympy 
tenacity==9.1.2
terminado 
tinycss2 
tokenizers 
tomli 
torch==2.6.0
tornado 
tqdm 
traitlets 
transformers 
typer==0.16.0
typer-slim==0.16.0
types-python-dateutil 
typing-inspection 
typing_extensions 
typing_utils 
uri-template 
urllib3 
uvicorn 
watchfiles 
wcwidth 
webcolors 
webencodings 
websocket-client 
websockets
widgetsnbextension 
win-inet-pton 
wrapt
xxhash 
yarl 
zipp
zstandard==0.23.0
